// src/API.ts
export type CreateMessageInput = {
  id?: string | null,
  clientId?: string | null,
  createdAt?: string | null,
  type: MessageType,
  message: string,
  name: string,
};

export enum MessageType {
  MESSAGE = "MESSAGE",
}


export type ModelMessageConditionInput = {
  clientId?: ModelIDInput | null,
  createdAt?: ModelStringInput | null,
  type?: ModelMessageTypeInput | null,
  message?: ModelStringInput | null,
  name?: ModelStringInput | null,
  and?: Array< ModelMessageConditionInput | null > | null,
  or?: Array< ModelMessageConditionInput | null > | null,
  not?: ModelMessageConditionInput | null,
};

export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export type ModelMessageTypeInput = {
  eq?: MessageType | null,
  ne?: MessageType | null,
};

export type Message = {
  __typename: "Message",
  id: string,
  clientId?: string | null,
  createdAt: string,
  type: MessageType,
  message: string,
  name: string,
  updatedAt: string,
};

export type UpdateMessageInput = {
  id: string,
  clientId?: string | null,
  createdAt?: string | null,
  type?: MessageType | null,
  message?: string | null,
  name?: string | null,
};

export type DeleteMessageInput = {
  id: string,
};

export type ModelMessageFilterInput = {
  id?: ModelIDInput | null,
  clientId?: ModelIDInput | null,
  createdAt?: ModelStringInput | null,
  type?: ModelMessageTypeInput | null,
  message?: ModelStringInput | null,
  name?: ModelStringInput | null,
  and?: Array< ModelMessageFilterInput | null > | null,
  or?: Array< ModelMessageFilterInput | null > | null,
  not?: ModelMessageFilterInput | null,
};

export type ModelMessageConnection = {
  __typename: "ModelMessageConnection",
  items:  Array<Message | null >,
  nextToken?: string | null,
};

export type ModelStringKeyConditionInput = {
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
};

export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}
// ...
export type CreateMessageMutation = {
  createMessage?:  {
    __typename: "Message",
    id: string,
    clientId?: string | null,
    createdAt: string,
    type: MessageType,
    message: string,
    name: string,
    updatedAt: string,
  } | null,
};
// ...
export type MessagesByDateQueryVariables = {
  type: MessageType,
  createdAt?: ModelStringKeyConditionInput | null,
  sortDirection?: ModelSortDirection | null,
  filter?: ModelMessageFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type MessagesByDateQuery = {
  messagesByDate?:  {
    __typename: "ModelMessageConnection",
    items:  Array< {
      __typename: "Message",
      id: string,
      clientId?: string | null,
      createdAt: string,
      type: MessageType,
      message: string,
      name: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type OnCreateMessageSubscriptionVariables = {
  filter?: ModelSubscriptionMessageFilterInput | null,
};

export type OnCreateMessageSubscription = {
  onCreateMessage?:  {
    __typename: "Message",
    id: string,
    clientId?: string | null,
    createdAt: string,
    type: MessageType,
    message: string,
    name: string,
    updatedAt: string,
  } | null,
};